+++ src/main/java/com/github/mjkuranda/spaceadventure2/GameData.java
@@ -7,10 +7,13 @@ import com.github.mjkuranda.spaceadventure2.entities.missiles.MissileType;
 import com.github.mjkuranda.spaceadventure2.entities.Particle;
 import com.github.mjkuranda.spaceadventure2.resources.GameAnimation;
 import com.github.mjkuranda.spaceadventure2.states.StatesId;
+import com.github.mjkuranda.spaceadventure2.states.highscore.HighScoreHandler;
+import com.github.mjkuranda.spaceadventure2.states.highscore.HighScoreRecord;
 import org.newdawn.slick.GameContainer;
 import org.newdawn.slick.Input;
 import org.newdawn.slick.state.StateBasedGame;
 
+import java.time.LocalDate;
 import java.util.*;
 
 public class GameData {
@@ -147,6 +150,9 @@ public class GameData {
             e.remove();
 
             if (!player.isAlive()) {
+                PlayerStatistics stats = PlayerStatistics.getInstance();
+                HighScoreHandler.getInstance().inputRecord(new HighScoreRecord(stats.getName(), stats.getScore(), LocalDate.now().toString()));
+
                 game.enterState(StatesId.GAME_OVER_MENU);
                 in.clearKeyPressedRecord();
                 reset();