+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/MenuState.java
@@ -10,9 +10,13 @@ import org.newdawn.slick.state.StateBasedGame;
 
 public abstract class MenuState extends BasicGameState {
 
-    private MenuOption[] options;
+    protected MenuOption[] options;
 
-    private int currentOption;
+    protected int currentOption;
+
+    public MenuState() {
+        this(null);
+    }
 
     public MenuState(MenuOption[] options) {
         this.options = options;