+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - 2023-08-05
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): HighScore state.
+- [@mjkuranda](https://github.com/mjkuranda): HighScoreRecord and its DTO.
+- [@mjkuranda](https://github.com/mjkuranda): HighScoreHandler to handle high score.
+- [@mjkuranda](https://github.com/mjkuranda): Saving and writing to the file of high scores.
+
+### Fixed
+- [@mjkuranda](https://github.com/mjkuranda): Resetting the other menu states after entering to another one.
+
 ## [0.17.0] - 2023-08-03
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Loading state.